refactor(indexer/rpc): extract client + logging + rate-limit to rpc/client.ts (PR-S6)
Move three blocks (~254 lines) out of indexer-envio/src/rpc.ts (1882 →
1628 lines) into the new indexer-envio/src/rpc/client.ts (289 lines):

- RPC failure logging (lines 21-110): _rpcFailureCounts map,
  sanitizeErrorMessage, KNOWN_REVERT_SIGNATURES, extractRevertSignature,
  logRpcFailure
- RPC client management (lines 253-433): rpcClients/fallbackRpcClients
  caches, _clearRpcClients, _setRpcClientForTests, RPC_CONFIG_BY_CHAIN,
  withHyperRpcToken, isBareHyperRpcUrl, getRpcClient, getFallbackRpcClient
- Rate-limit detection (lines 435-450): RATE_LIMIT_RE, isRateLimitError,
  RATE_LIMIT_RETRY_DELAYS_MS

The previously-exported public surface (getRpcClient, withHyperRpcToken,
_clearRpcClients, _setRpcClientForTests) is re-exported from rpc.ts so
external callers (feeToken.ts, EventHandlers.ts, breakers.ts,
hyperRpcToken.test.ts, blockFallback.test.ts, rpcClient.test.ts) keep
working without import-path changes. The fetchers that remain in rpc.ts
(fetchReserves, fetchTradingLimits, fetchFees, fetchBreakerKind, etc.)
import everything they need from "./rpc/client".

Pinned by PR-S5's 24-test characterization suite — `pnpm test` reports
381 passing.
